pyproject.toml
[project]
name = "agent-lab"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "openai>=2.8.1",
    "pydantic>=2.12.4",
    "python-dotenv>=1.2.1",
    "rich>=14.2.0",
    "typer>=0.20.0",
]
